+++ lib/view/profile/profile_screen.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:simple_notes_app/view/authentication/authentication.dart';
 import 'package:simple_notes_app/view/profile/profile_screen_widgets.dart';
+import 'package:simple_notes_app/view/profile/widgets/profile_card.dart';
 import 'package:simple_notes_app/widgets_library/widgets_library.dart';
 
 class ProfileScreen extends StatelessWidget {
@@ -18,7 +21,14 @@ class ProfileScreen extends StatelessWidget {
             sliver: SliverList(
               delegate: SliverChildListDelegate(
                 [
-                  const ProfileCard(),
+                  BlocBuilder<AuthBloc, AuthState>(
+                    builder: (context, state) {
+                      return state.maybeWhen(
+                        orElse: () => const ProfileCard(user: null),
+                        signedIn: (user) => ProfileCard(user: user),
+                      );
+                    },
+                  ),
                   const SizedBox(height: 30),
                   const ItemTilesArea(),
                 ],
