+++ lib/view/profile/profile_screen_widgets.dart
@@ -1,61 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:simple_notes_app/core/extensions.dart';
 
-class ProfileCard extends StatelessWidget {
-  const ProfileCard({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(10),
-      ),
-      child: Padding(
-        padding: const EdgeInsets.symmetric(
-          horizontal: 8,
-          vertical: 32,
-        ),
-        child: Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            const CircleAvatar(
-              child: Icon(
-                Icons.account_circle_rounded,
-              ),
-            ),
-            const SizedBox(width: 10),
-            Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                TextButton(
-                  style: TextButton.styleFrom(
-                    padding: EdgeInsets.zero,
-                    minimumSize: Size.zero,
-                    tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                  ),
-                  onPressed: () {
-                    context.push('/auth/sign-in');
-                  },
-                  child: Text(
-                    'Sign In',
-                    style: context.theme.textButtonTheme.style?.textStyle
-                        ?.resolve({
-                      ...MaterialState.values,
-                    })?.copyWith(
-                      decoration: TextDecoration.underline,
-                    ),
-                  ),
-                ),
-                const Text('Back up, Sync'),
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
 
 class ItemTilesArea extends StatelessWidget {
   const ItemTilesArea({
